ESP8266 works, really.
ESP8266 might not be supported, but it works. I mean - there is even big prominent "Examples - ESP8266" which links to [https://github.com/nothans/ESP8266](here) which is really bad.
This might be the example by supporting range of platforms. Oh, not *supporting* but still works.
But to let the spirit stay, ESP32 is still not supported (someone please check as I don't have one at hand).